Overlay refactoring and improvement
